Simple JSF 2 + Wildfly 10 + CDI demo CRUD web application (without database configuration)
This project aimed to show some good practices with/and using JSF in a simple CRUD structure.  It uses maven and pom.xml is configured with the current versions of each technology.

This simple project was based on this URL:
http://www.codeproject.com/Articles/1030872/A-Simple-CRUD-Example-with-JSF

After configuration of your IDE (IntelliJ IDEA in my case), you should be able to access using the common url:
http://localhost:8080/demo